[FIX] 백엔드 정리 및 최적화 작업
- SQLite NFS 환경 호환성 개선
- Pydantic 설정 최적화
- 테스트 파일 구조 개선 (tests/ 디렉터리로 통합)
- 프로젝트 구조 정리
- Kubernetes 배포 준비
