+++ tests/test_advanced_nlp_integration.py
@@ -1,420 +1,357 @@
 #!/usr/bin/env python3
 """
-고급 NLP 통합 테스트 스크립트
-다중 모델, 컨텍스트 인식, 지능적 해석, 학습 기반 개선 기능을 통합 테스트
+고급 자연어 처리 통합 테스트 스크립트
+T-047 고급 자연어 처리 및 AI 모델 통합 완료 테스트
 """
 
 import asyncio
-import pytest
-import json
-from unittest.mock import Mock, patch, AsyncMock
-from datetime import datetime
+import logging
+import sys
+import os
 from typing import Dict, Any
+from datetime import datetime
 
-# 테스트 대상 모듈들
+# 프로젝트 루트를 Python 경로에 추가
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'app'))
+
+from app.llm.gemini import GeminiClient
 from app.llm.advanced_nlp_service import AdvancedNLPService
-from app.llm.multi_model_processor import MultiModelProcessor, DEFAULT_CONFIG
-from app.llm.context_manager import ContextManager
-from app.llm.smart_command_interpreter import SmartCommandInterpreter
-from app.llm.learning_processor import LearningProcessor, LearningFeedback
-from app.mcp.tools.advanced_nlp import process_advanced_command, record_user_feedback, get_user_insights
+from app.core.config import get_settings
 
+# 로깅 설정
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+)
+logger = logging.getLogger(__name__)
 
-class TestAdvancedNLPIntegration:
+class AdvancedNLPIntegrationTester:
     """고급 NLP 통합 테스트 클래스"""
     
-    @pytest.fixture
-    async def mock_config(self):
-        """모의 설정"""
-        return {
-            "gemini": {
-                "enabled": True,
-                "base_url": "https://generativelanguage.googleapis.com",
-                "api_key": "test-api-key"
-            },
-            "claude": {
-                "enabled": False,
-                "base_url": "https://api.anthropic.com",
-                "api_key": "test-claude-key"
-            },
-            "gpt4": {
-                "enabled": False,
-                "base_url": "https://api.openai.com",
-                "api_key": "test-gpt4-key"
-            }
-        }
+    def __init__(self):
+        self.settings = get_settings()
+        self.gemini_client = GeminiClient()
+        self.advanced_nlp_service = None
+        
+    async def setup(self):
+        """테스트 환경 설정"""
+        try:
+            logger.info("고급 NLP 통합 테스트 환경 설정 중...")
+            
+            # 고급 NLP 서비스 초기화
+            self.advanced_nlp_service = AdvancedNLPService()
+            await self.advanced_nlp_service.initialize()
+            
+            logger.info("테스트 환경 설정 완료")
+            return True
+            
+        except Exception as e:
+            logger.error(f"테스트 환경 설정 실패: {e}")
+            return False
     
-    @pytest.fixture
-    async def mock_context(self):
-        """모의 컨텍스트"""
-        return {
-            "project_name": "test-project",
-            "current_deployments": [
-                {"name": "web-app", "environment": "staging", "replicas": 2},
-                {"name": "api-service", "environment": "production", "replicas": 3}
-            ],
-            "user_id": "test-user-123",
-            "recent_conversations": [
-                {
-                    "user_input": "앱 상태 확인해줘",
-                    "system_response": "현재 배포된 앱들의 상태를 확인했습니다",
-                    "action_taken": "status_check",
-                    "success": True
-                }
-            ]
-        }
-    
-    @pytest.fixture
-    async def advanced_nlp_service(self, mock_config):
-        """고급 NLP 서비스 인스턴스"""
-        service = AdvancedNLPService(mock_config)
-        await service.initialize()
-        return service
-    
-    @pytest.mark.asyncio
-    async def test_advanced_command_processing(self, advanced_nlp_service, mock_context):
-        """고급 명령 처리 테스트"""
-        # 테스트 명령
-        test_command = "web-app을 staging에 3개 복제본으로 배포해줘"
+    async def test_basic_gemini_integration(self):
+        """기본 Gemini 통합 테스트"""
+        logger.info("=== 기본 Gemini 통합 테스트 ===")
         
-        # Gemini API 모의 응답
-        mock_gemini_response = {
-            "candidates": [{
-                "content": {
-                    "parts": [{
-                        "text": json.dumps({
-                            "action": "deploy",
-                            "target": "web-app",
-                            "parameters": {
-                                "replicas": 3,
-                                "environment": "staging"
-                            },
-                            "confidence": 0.95,
-                            "ambiguities": [],
-                            "suggestions": [],
-                            "alternatives": [],
-                            "warnings": [],
-                            "next_steps": ["배포 완료 후 상태 확인"],
-                            "learning_notes": "복제본 수와 환경을 명확히 지정한 좋은 명령입니다"
-                        })
-                    }]
-                }
-            }],
-            "usageMetadata": {
-                "totalTokenCount": 150
-            }
-        }
-        
-        with patch('httpx.AsyncClient.post') as mock_post:
-            mock_response = Mock()
-            mock_response.json.return_value = mock_gemini_response
-            mock_response.raise_for_status.return_value = None
-            mock_post.return_value = mock_response
-            
-            # 명령 처리 실행
-            result = await advanced_nlp_service.process_command(
-                user_id="test-user-123",
-                project_name="test-project",
-                command=test_command,
-                context=mock_context
-            )
-            
-            # 결과 검증
-            assert result["original_command"] == test_command
-            assert result["user_id"] == "test-user-123"
-            assert result["project_name"] == "test-project"
-            assert "interpreted_command" in result
-            assert "model_responses" in result
-            assert "ambiguities" in result
-            assert "suggestions" in result
-            assert "learned_suggestions" in result
-            assert "processing_metadata" in result
-            
-            # 처리 메타데이터 검증
-            metadata = result["processing_metadata"]
-            assert metadata["models_used"] > 0
-            assert metadata["total_processing_time"] > 0
-    
-    @pytest.mark.asyncio
-    async def test_context_aware_processing(self, mock_config, mock_context):
-        """컨텍스트 인식 처리 테스트"""
-        # 컨텍스트 관리자 모의
-        context_manager = ContextManager("redis://localhost:6379")
-        context_manager.redis_client = None  # 메모리 모드
-        context_manager._memory_store = {}
-        
-        # 다중 모델 프로세서 모의
-        multi_processor = MultiModelProcessor(mock_config)
-        
-        # 컨텍스트 인식 프로세서 생성
-        from app.llm.context_manager import ContextAwareProcessor
-        processor = ContextAwareProcessor(context_manager, multi_processor)
-        
-        # Gemini API 모의
-        mock_gemini_response = {
-            "candidates": [{
-                "content": {
-                    "parts": [{
-                        "text": json.dumps({
-                            "action": "deploy",
-                            "target": "web-app",
-                            "parameters": {"replicas": 3, "environment": "staging"},
-                            "confidence": 0.9
-                        })
-                    }]
-                }
-            }],
-            "usageMetadata": {"totalTokenCount": 100}
-        }
-        
-        with patch('httpx.AsyncClient.post') as mock_post:
-            mock_response = Mock()
-            mock_response.json.return_value = mock_gemini_response
-            mock_response.raise_for_status.return_value = None
-            mock_post.return_value = mock_response
-            
-            # 컨텍스트 인식 처리 실행
-            result = await processor.process_with_context(
-                user_id="test-user-123",
-                project_name="test-project",
-                command="web-app 배포해줘"
-            )
-            
-            # 결과 검증
-            assert "processed_command" in result
-            assert "context" in result
-            assert "conversation_turn" in result
-            
-            # 컨텍스트 검증
-            context = result["context"]
-            assert context["user_id"] == "test-user-123"
-            assert context["project_name"] == "test-project"
-            assert "recent_conversations" in context
-    
-    @pytest.mark.asyncio
-    async def test_smart_command_interpretation(self):
-        """지능적 명령 해석 테스트"""
-        interpreter = SmartCommandInterpreter()
-        
-        # 테스트 케이스들
-        test_cases = [
-            {
-                "command": "앱 배포해줘",
-                "expected_ambiguities": ["MISSING_TARGET"],
-                "expected_suggestions": True
-            },
-            {
-                "command": "my-app을 staging에 3개로 배포해줘",
-                "expected_ambiguities": [],
-                "expected_suggestions": False
-            },
-            {
-                "command": "스케일링",
-                "expected_ambiguities": ["MISSING_TARGET", "MISSING_PARAMETERS"],
-                "expected_suggestions": True
-            }
+        test_commands = [
+            "myapp을 스테이징에 배포해줘",
+            "testapp을 프로덕션에 롤백해줘",
+            "myapp의 상태를 확인해줘",
+            "testapp을 3개로 스케일링해줘"
         ]
         
-        context = {
-            "current_deployments": [
-                {"name": "my-web-app", "environment": "staging"},
-                {"name": "my-api", "environment": "production"}
-            ]
-        }
-        
-        for test_case in test_cases:
-            result = await interpreter.interpret_command(
-                test_case["command"], 
-                context
-            )
-            
-            # 기본 검증
-            assert result.original_command == test_case["command"]
-            assert result.confidence >= 0.0
-            assert result.confidence <= 1.0
-            
-            # 모호함 검증
-            if test_case["expected_ambiguities"]:
-                assert len(result.ambiguities) > 0
-                ambiguity_types = [amb.ambiguity_type.value for amb in result.ambiguities]
-                for expected_type in test_case["expected_ambiguities"]:
-                    assert expected_type in ambiguity_types
-            
-            # 제안 검증
-            if test_case["expected_suggestions"]:
-                assert len(result.suggestions) > 0
-            else:
-                assert len(result.suggestions) == 0
-    
-    @pytest.mark.asyncio
-    async def test_learning_processor(self):
-        """학습 프로세서 테스트"""
-        learning_processor = LearningProcessor("redis://localhost:6379")
-        learning_processor.redis_client = None  # 메모리 모드
-        learning_processor._memory_store = {}
-        
-        # 피드백 생성
-        feedback = LearningFeedback(
-            user_id="test-user-123",
-            command="앱 배포해줘",
-            original_interpretation={
-                "action": "deploy",
-                "target": "unknown",
-                "confidence": 0.6,
-                "model_used": "gemini"
-            },
-            user_correction={
-                "action": "deploy",
-                "target": "my-web-app",
-                "environment": "staging"
-            },
-            feedback_type="correction",
-            timestamp=datetime.now(),
-            success=True,
-            context={"project_name": "test-project"}
-        )
-        
-        # 피드백 기록
-        await learning_processor.record_feedback(feedback)
-        
-        # 학습된 제안 조회
-        suggestions = await learning_processor.get_learned_suggestions(
-            "앱 배포해줘",
-            "test-user-123",
-            {"project_name": "test-project"}
-        )
-        
-        # 결과 검증
-        assert isinstance(suggestions, list)
-        # 메모리 모드에서는 패턴이 저장되지 않을 수 있음
-        # assert len(suggestions) >= 0
-    
-    @pytest.mark.asyncio
-    async def test_mcp_tools_integration(self, mock_config):
-        """MCP 도구 통합 테스트"""
-        # Gemini API 모의
-        mock_gemini_response = {
-            "candidates": [{
-                "content": {
-                    "parts": [{
-                        "text": json.dumps({
-                            "action": "deploy",
-                            "target": "test-app",
-                            "parameters": {"replicas": 2, "environment": "staging"},
-                            "confidence": 0.9
-                        })
-                    }]
-                }
-            }],
-            "usageMetadata": {"totalTokenCount": 120}
-        }
-        
-        with patch('httpx.AsyncClient.post') as mock_post:
-            mock_response = Mock()
-            mock_response.json.return_value = mock_gemini_response
-            mock_response.raise_for_status.return_value = None
-            mock_post.return_value = mock_response
-            
-            # MCP 도구 테스트
-            result = await process_advanced_command(
-                user_id="test-user-123",
-                project_name="test-project",
-                command="test-app 배포해줘"
-            )
-            
-            # 결과 검증
-            assert result["original_command"] == "test-app 배포해줘"
-            assert result["user_id"] == "test-user-123"
-            assert result["project_name"] == "test-project"
-            assert "model_response" in result
-            assert "advanced_features" in result
-    
-    @pytest.mark.asyncio
-    async def test_error_handling(self, advanced_nlp_service, mock_context):
-        """에러 처리 테스트"""
-        # API 에러 시뮬레이션
-        with patch('httpx.AsyncClient.post') as mock_post:
-            mock_post.side_effect = Exception("API Error")
-            
-            # 에러가 발생해도 서비스가 중단되지 않는지 확인
+        results = []
+        for command in test_commands:
             try:
-                result = await advanced_nlp_service.process_command(
-                    user_id="test-user-123",
-                    project_name="test-project",
-                    command="test command",
-                    context=mock_context
+                logger.info(f"테스트 명령: {command}")
+                result = await self.gemini_client.interpret(
+                    prompt=command,
+                    user_id="test_user",
+                    project_name="test_project"
                 )
-                # 에러가 발생하면 빈 결과가 반환되어야 함
-                assert result is not None
+                
+                results.append({
+                    "command": command,
+                    "success": "error" not in result,
+                    "intent": result.get("intent", "unknown"),
+                    "message": result.get("message", ""),
+                    "llm_mode": result.get("llm", {}).get("mode", "unknown")
+                })
+                
+                logger.info(f"결과: {result.get('intent', 'unknown')} - {result.get('message', '')}")
+                
             except Exception as e:
-                # 예상된 에러인지 확인
-                assert "API Error" in str(e)
-    
-    @pytest.mark.asyncio
-    async def test_performance_metrics(self, advanced_nlp_service, mock_context):
-        """성능 메트릭 테스트"""
-        # Gemini API 모의
-        mock_gemini_response = {
-            "candidates": [{
-                "content": {
-                    "parts": [{
-                        "text": json.dumps({
-                            "action": "deploy",
-                            "target": "test-app",
-                            "confidence": 0.9
-                        })
-                    }]
-                }
-            }],
-            "usageMetadata": {"totalTokenCount": 100}
-        }
+                logger.error(f"명령 처리 실패: {command} - {e}")
+                results.append({
+                    "command": command,
+                    "success": False,
+                    "error": str(e)
+                })
         
-        with patch('httpx.AsyncClient.post') as mock_post:
-            mock_response = Mock()
-            mock_response.json.return_value = mock_gemini_response
-            mock_response.raise_for_status.return_value = None
-            mock_post.return_value = mock_response
-            
-            # 성능 측정
-            start_time = asyncio.get_event_loop().time()
-            
-            result = await advanced_nlp_service.process_command(
-                user_id="test-user-123",
-                project_name="test-project",
-                command="test-app 배포해줘",
-                context=mock_context
+        return results
+    
+    async def test_advanced_nlp_processing(self):
+        """고급 NLP 처리 테스트"""
+        logger.info("=== 고급 NLP 처리 테스트 ===")
+        
+        if not self.advanced_nlp_service:
+            logger.error("고급 NLP 서비스가 초기화되지 않았습니다.")
+            return []
+        
+        test_commands = [
+            "앱을 배포해줘",  # 모호한 명령
+            "my-web-app을 staging 환경에 2개 인스턴스로 배포해줘",  # 명확한 명령
+            "이전 버전으로 되돌려줘",  # 컨텍스트가 필요한 명령
+            "서버 상태 확인",  # 간단한 명령
+        ]
+        
+        results = []
+        for command in test_commands:
+            try:
+                logger.info(f"고급 NLP 테스트 명령: {command}")
+                result = await self.advanced_nlp_service.process_command(
+                    user_id="test_user",
+                    project_name="test_project",
+                    command=command,
+                    context={"current_deployments": [{"name": "my-web-app"}]}
+                )
+                
+                results.append({
+                    "command": command,
+                    "success": True,
+                    "confidence": result.get("confidence", 0.0),
+                    "quality": result.get("quality", "unknown"),
+                    "best_model": result.get("best_model", "unknown"),
+                    "ambiguities_count": len(result.get("ambiguities", [])),
+                    "suggestions_count": len(result.get("suggestions", [])),
+                    "learned_suggestions_count": len(result.get("learned_suggestions", []))
+                })
+                
+                logger.info(f"신뢰도: {result.get('confidence', 0.0):.2f}, "
+                          f"품질: {result.get('quality', 'unknown')}, "
+                          f"모델: {result.get('best_model', 'unknown')}")
+                
+            except Exception as e:
+                logger.error(f"고급 NLP 처리 실패: {command} - {e}")
+                results.append({
+                    "command": command,
+                    "success": False,
+                    "error": str(e)
+                })
+        
+        return results
+    
+    async def test_feedback_learning(self):
+        """피드백 학습 테스트"""
+        logger.info("=== 피드백 학습 테스트 ===")
+        
+        if not self.advanced_nlp_service:
+            logger.error("고급 NLP 서비스가 초기화되지 않았습니다.")
+            return False
+        
+        try:
+            # 테스트 명령 처리
+            command = "앱을 배포해줘"
+            result = await self.advanced_nlp_service.process_command(
+                user_id="test_user",
+                project_name="test_project",
+                command=command
             )
             
-            end_time = asyncio.get_event_loop().time()
-            total_time = end_time - start_time
+            # 피드백 기록
+            await self.advanced_nlp_service.record_feedback(
+                user_id="test_user",
+                command=command,
+                original_interpretation=result["interpreted_command"],
+                user_correction={
+                    "action": "deploy",
+                    "target": "my-web-app",
+                    "environment": "staging"
+                },
+                feedback_type="correction",
+                success=True
+            )
             
-            # 성능 검증
-            assert total_time < 5.0  # 5초 이내 완료
-            assert result["processing_metadata"]["total_processing_time"] > 0
-            assert result["processing_metadata"]["models_used"] > 0
-
-
-# 통합 테스트 실행 함수
-async def run_integration_tests():
-    """통합 테스트 실행"""
-    print("=== 고급 NLP 통합 테스트 시작 ===")
+            logger.info("피드백 학습 테스트 완료")
+            return True
+            
+        except Exception as e:
+            logger.error(f"피드백 학습 테스트 실패: {e}")
+            return False
     
-    # 테스트 실행
-    pytest.main([
-        __file__,
-        "-v",
-        "--tb=short",
-        "--asyncio-mode=auto"
-    ])
+    async def test_user_insights(self):
+        """사용자 인사이트 테스트"""
+        logger.info("=== 사용자 인사이트 테스트 ===")
+        
+        if not self.advanced_nlp_service:
+            logger.error("고급 NLP 서비스가 초기화되지 않았습니다.")
+            return None
+        
+        try:
+            insights = await self.advanced_nlp_service.get_user_insights("test_user")
+            
+            logger.info(f"사용자 인사이트 조회 완료:")
+            logger.info(f"- 최근 대화: {insights.get('recent_conversations', 0)}개")
+            logger.info(f"- 성공률: {insights.get('user_pattern', {}).get('success_rate', 0.0):.1%}")
+            logger.info(f"- 학습된 제안: {len(insights.get('learned_suggestions', []))}개")
+            
+            return insights
+            
+        except Exception as e:
+            logger.error(f"사용자 인사이트 테스트 실패: {e}")
+            return None
     
-    print("=== 고급 NLP 통합 테스트 완료 ===")
+    async def test_error_handling(self):
+        """오류 처리 테스트"""
+        logger.info("=== 오류 처리 테스트 ===")
+        
+        error_commands = [
+            "",  # 빈 명령
+            "invalid_command_xyz",  # 잘못된 명령
+            None,  # None 값
+        ]
+        
+        results = []
+        for command in error_commands:
+            try:
+                if command is None:
+                    # None 값 테스트
+                    result = await self.gemini_client.interpret(
+                        prompt=None,  # type: ignore
+                        user_id="test_user",
+                        project_name="test_project"
+                    )
+                else:
+                    result = await self.gemini_client.interpret(
+                        prompt=command,
+                        user_id="test_user",
+                        project_name="test_project"
+                    )
+                
+                results.append({
+                    "command": str(command),
+                    "success": "error" in result,
+                    "error_handled": True
+                })
+                
+            except Exception as e:
+                results.append({
+                    "command": str(command),
+                    "success": False,
+                    "error_handled": True,
+                    "exception": str(e)
+                })
+        
+        return results
+    
+    async def run_all_tests(self):
+        """모든 테스트 실행"""
+        logger.info("고급 NLP 통합 테스트 시작")
+        logger.info("=" * 50)
+        
+        # 환경 설정
+        if not await self.setup():
+            logger.error("테스트 환경 설정 실패")
+            return False
+        
+        test_results = {
+            "timestamp": datetime.now().isoformat(),
+            "basic_gemini": [],
+            "advanced_nlp": [],
+            "feedback_learning": False,
+            "user_insights": None,
+            "error_handling": [],
+            "overall_success": False
+        }
+        
+        try:
+            # 1. 기본 Gemini 통합 테스트
+            test_results["basic_gemini"] = await self.test_basic_gemini_integration()
+            
+            # 2. 고급 NLP 처리 테스트
+            test_results["advanced_nlp"] = await self.test_advanced_nlp_processing()
+            
+            # 3. 피드백 학습 테스트
+            test_results["feedback_learning"] = await self.test_feedback_learning()
+            
+            # 4. 사용자 인사이트 테스트
+            test_results["user_insights"] = await self.test_user_insights()
+            
+            # 5. 오류 처리 테스트
+            test_results["error_handling"] = await self.test_error_handling()
+            
+            # 전체 성공 여부 계산
+            basic_success = all(r.get("success", False) for r in test_results["basic_gemini"])
+            advanced_success = all(r.get("success", False) for r in test_results["advanced_nlp"])
+            feedback_success = test_results["feedback_learning"]
+            error_success = all(r.get("error_handled", False) for r in test_results["error_handling"])
+            
+            test_results["overall_success"] = (
+                basic_success and advanced_success and 
+                feedback_success and error_success
+            )
+            
+            # 결과 출력
+            self.print_test_results(test_results)
+            
+            return test_results["overall_success"]
+            
+        except Exception as e:
+            logger.error(f"테스트 실행 중 오류 발생: {e}")
+            return False
+        
+        finally:
+            # 정리
+            if self.advanced_nlp_service:
+                await self.advanced_nlp_service.close()
+    
+    def print_test_results(self, results: Dict[str, Any]):
+        """테스트 결과 출력"""
+        logger.info("=" * 50)
+        logger.info("테스트 결과 요약")
+        logger.info("=" * 50)
+        
+        # 기본 Gemini 테스트 결과
+        basic_results = results["basic_gemini"]
+        basic_success = sum(1 for r in basic_results if r.get("success", False))
+        logger.info(f"기본 Gemini 통합: {basic_success}/{len(basic_results)} 성공")
+        
+        # 고급 NLP 테스트 결과
+        advanced_results = results["advanced_nlp"]
+        advanced_success = sum(1 for r in advanced_results if r.get("success", False))
+        logger.info(f"고급 NLP 처리: {advanced_success}/{len(advanced_results)} 성공")
+        
+        # 피드백 학습 결과
+        logger.info(f"피드백 학습: {'성공' if results['feedback_learning'] else '실패'}")
+        
+        # 사용자 인사이트 결과
+        insights = results["user_insights"]
+        if insights and "error" not in insights:
+            logger.info(f"사용자 인사이트: 성공 (대화 {insights.get('recent_conversations', 0)}개)")
+        else:
+            logger.info("사용자 인사이트: 실패")
+        
+        # 오류 처리 결과
+        error_results = results["error_handling"]
+        error_success = sum(1 for r in error_results if r.get("error_handled", False))
+        logger.info(f"오류 처리: {error_success}/{len(error_results)} 성공")
+        
+        # 전체 결과
+        logger.info("=" * 50)
+        logger.info(f"전체 테스트 결과: {'성공' if results['overall_success'] else '실패'}")
+        logger.info("=" * 50)
 
+async def main():
+    """메인 함수"""
+    tester = AdvancedNLPIntegrationTester()
+    success = await tester.run_all_tests()
+    
+    if success:
+        logger.info("모든 테스트가 성공적으로 완료되었습니다!")
+        sys.exit(0)
+    else:
+        logger.error("일부 테스트가 실패했습니다.")
+        sys.exit(1)
 
 if __name__ == "__main__":
-    asyncio.run(run_integration_tests())
-
-
-
-
-
-
+    asyncio.run(main())
