+++ tests/test_cicd_mcp_trigger.py
