+++ tests/test_mcp_circuit_breaker.py
