+++ tests/test_mcp_endpoints.py
