+++ tests/test_mcp_external.py
