+++ tests/test_slack_templates.py
